requirements.txt
annotated-types==0.7.0
anyio==4.8.0
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
distro==1.9.0
exceptiongroup==1.2.2
Flask==3.1.0
google-search-results==2.4.2
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.8.2
MarkupSafe==3.0.2
openai==1.65.4
pinecone==6.0.1
pinecone-plugin-interface==0.0.7
pydantic==2.10.6
pydantic_core==2.27.2
pyTelegramBotAPI==4.26.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-telegram-bot==21.11.1
requests==2.32.3
six==1.17.0
sniffio==1.3.1
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3

beautifulsoup4~=4.13.3
soupsieve~=2.6
packaging~=24.2
bleach~=6.2.0
defusedxml~=0.7.1
ipython~=8.31.0
numpy~=2.0.2
pillow~=11.1.0
pip~=23.2.1
attrs~=24.3.0
wheel~=0.41.2
PySocks~=1.7.1
protobuf~=5.29.3
tornado~=6.4.2
scipy~=1.15.2
scikit-learn~=1.6.1
rich~=13.9.4
keras~=3.9.0
colorama~=0.4.6
notebook~=7.3.2
PyYAML~=6.0.2
Pygments~=2.19.1
joblib~=1.4.2
lz4~=4.4.3
psutil~=6.1.1
pandas~=2.2.3
wcwidth~=0.2.13
telebot~=0.0.5
grpcio~=1.70.0
pytz~=2025.1
setuptools~=68.2.0
filelock~=3.17.0
dotenv~=0.9.9
pytesseract~=0.3.13
pytest~=8.3.5
pluggy~=1.5.0
iniconfig~=2.0.0
django-googlesearch~=0.2
lxml~=5.3.1
jupyterlab-pygments~=0.3.0
regex~=2024.11.6
sentence-transformers~=3.4.1
bs4~=0.0.2
nltk~=3.9.1